requirements.txt
absl-py==2.2.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
grpcio==1.71.0
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scipy==1.15.2
setuptools==77.0.3
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torch-tb-profiler==0.4.3
torchinfo==1.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
